% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{comp_delta}
\alias{comp_delta}
\title{Expected Hellinger distance matrix}
\usage{
comp_delta(theta, p, n)
}
\arguments{
\item{theta}{List of samples from the localized atoms in a GMM. Each item in the list is an MCMC sample of the localized atoms for the data.}

\item{p}{Dimension of the data.}

\item{n}{Sample size, number of objects to cluster.}
}
\value{
An \code{n}x\code{n} symmetric matrix of the pairwise expected Hellinger distance for all objects.
}
\description{
Computes the posterior expected Hellinger distance matrix for all localized densities.
}
