% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow.R
\name{elbow}
\alias{elbow}
\title{FOLD elbow plot}
\usage{
elbow(tree, Delta)
}
\arguments{
\item{tree}{Matrix of candidate clusterings generated by hierarchical clustering on Delta.
Each row index is the number of clusters.}

\item{Delta}{An nxn symmetric matrix of pairwise Hellinger distances.}
}
\value{
An elbow plot showing the average Hellinger separation for each candidate clustering.
}
\description{
Creates the elbow plot for choosing a clustering point estimate with FOLD.
}
\details{
The elbow plot computes the total within-cluster pairwise Hellinger distances for each candidate clustering. A heuristic for
choosing the number of clusters is to identify the candidate at which the plot bends.
}
