% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{fold_risk}
\alias{fold_risk}
\title{FOLD risk function}
\usage{
fold_risk(c, Delta, omega)
}
\arguments{
\item{c}{A vector of cluster labels with length \code{n}.}

\item{Delta}{An \code{n}x\code{n} symmetric matrix of pairwise expected Hellinger distances.}

\item{omega}{FOLD separation parameter, positive.}
}
\value{
The risk of the supplied clustering.
}
\description{
Computes the risk of any clustering according to the FOLD loss.
}
